[Modify] move `NullLogger` to testUtility.ts
