+++ src/base/common/logger.ts
@@ -198,21 +198,6 @@ export function parseToLogLevel(str?: string): LogLevel {
 }
 
 /**
- * @class A logger that does nothing. Usually used for testing purpose.
- */
-export class NullLogger extends AbstractLogger implements ILogService {
-    constructor() {
-        super();
-    }
-    public trace(message: string, ...args: any[]): void {}
-    public debug(message: string, ...args: any[]): void {}
-    public info(message: string, ...args: any[]): void {}
-    public warn(message: string, ...args: any[]): void {}
-    public error(message: string | Error, ...args: any[]): void {}
-    public fatal(message: string | Error, ...args: any[]): void {}
-}
-
-/**
  * @class A simple integrated {@link ILogger} that combines the other loggers
  * into a intergrated version.
  */
