+++ test/code/platform/abstractConfigService.test.ts
@@ -1,13 +1,12 @@
 import * as assert from 'assert';
 import { URI } from 'src/base/common/file/uri';
-import { NullLogger } from 'src/base/common/logger';
 import { AbstractConfigService } from 'src/code/platform/configuration/common/abstractConfigService';
 import { ConfigCollection } from 'src/code/platform/configuration/common/configCollection';
 import { BuiltInConfigScope, IConfigRegistrant } from 'src/code/platform/configuration/common/configRegistrant';
 import { DefaultConfigStorage } from 'src/code/platform/configuration/common/configStorage';
 import { FileService } from 'src/code/platform/files/common/fileService';
 import { Registrants } from 'src/code/platform/registrant/common/registrant';
-import { NullLifecycleService } from 'test/testUtility';
+import { NullLifecycleService, NullLogger } from 'test/testUtility';
 
 class TestDefaultConfigStorage extends DefaultConfigStorage {
     protected override createDefaultModel(): Record<PropertyKey, any> {
