+++ test/code/platform/abstractLifecycleService.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert';
-import { NullLogger } from 'src/base/common/logger';
 import { delayFor } from 'src/base/common/util/async';
 import { AbstractLifecycleService } from 'src/code/platform/lifecycle/common/abstractLifecycleService';
+import { NullLogger } from 'test/testUtility';
 
 suite('abstract-lifecycle-service-test', () => {
