+++ test/code/platform/diskStorage.test.ts
@@ -3,11 +3,11 @@ import * as fs from 'fs';
 import { afterEach, setup } from 'mocha';
 import { join } from 'src/base/common/file/path';
 import { Schemas, URI } from 'src/base/common/file/uri';
-import { ILogService, NullLogger } from 'src/base/common/logger';
+import { ILogService } from 'src/base/common/logger';
 import { DiskFileSystemProvider } from 'src/code/platform/files/node/diskFileSystemProvider';
 import { FileService, IFileService } from 'src/code/platform/files/common/fileService';
 import { DiskStorage } from 'src/code/platform/files/common/diskStorage';
-import { TestDir } from 'test/testUtility';
+import { TestDir, NullLogger } from 'test/testUtility';
 
 suite('storage-test', () => {
