+++ test/code/platform/i18n.test.ts
@@ -1,8 +1,8 @@
 import * as assert from 'assert';
-import { NullLogger } from 'src/base/common/logger';
 import { FileService, IFileService } from 'src/code/platform/files/common/fileService';
 import { i18n } from 'src/code/platform/i18n/i18n';
 import { Section } from 'src/code/platform/section';
+import { NullLogger } from 'test/testUtility';
 
 class i18nTest extends i18n {
