+++ test/testUtility.ts
@@ -1,6 +1,6 @@
 import { tmpdir } from "os";
 import { join } from "path";
-import { NullLogger } from "src/base/common/logger";
+import { AbstractLogger, ILogService } from "src/base/common/logger";
 import { DiskEnvironmentService } from "src/code/platform/environment/common/diskEnvironmentService";
 import { AbstractLifecycleService } from "src/code/platform/lifecycle/common/abstractLifecycleService";
 
@@ -31,4 +31,19 @@ export class NullEnvironmentService extends DiskEnvironmentService {
             },
         );
     }
+}
+
+/**
+ * @class A logger that does nothing. Usually used for testing purpose.
+ */
+export class NullLogger extends AbstractLogger implements ILogService {
+    constructor() {
+        super();
+    }
+    public trace(message: string, ...args: any[]): void {}
+    public debug(message: string, ...args: any[]): void {}
+    public info(message: string, ...args: any[]): void {}
+    public warn(message: string, ...args: any[]): void {}
+    public error(message: string | Error, ...args: any[]): void {}
+    public fatal(message: string | Error, ...args: any[]): void {}
 }
